_cursor_name_has_row_ is generated unconditionally, can yield warnings
Summary:
Do not generate the _cursor_has_row_ variable unless there is a FETCH INTO or the cursor is global.

To do this we added another flag on the cursor variable, which gives us an opportunity to check for cases where the cursor was declared, not ever fetched, and then its _has_row was used.

Reviewed By: winniequinn

Differential Revision: D35087481

fbshipit-source-id: 4726e6bd10cecc432c6f9926826247d7a4ea9a92