call foo(*) should allow you to pass matching locals as well as args
Summary:
the call foo(*) pattern expands to

```
call foo(from arguments like foo arguments);
```

but this is not quite what you really want, what you really want is:

```
call foo(from locals like foo arguments);
```

But there was no way to do that because there was no singular shape for `locals`.

This diff creates a synthetic locals shape instanteously when you use it.  It's
the declared unitary locals (not cursors) in order of declaration at the point
at which you use it.  The `*` pattern uses it now.

Reviewed By: winniequinn

Differential Revision: D35126965

fbshipit-source-id: 6e215a3c5ccf2fdbb9310f3795acbdbc91f75749