manage subscriptions like other entities to set up for JSON output
Summary:
We were using the schema annotations list to track all the subscriptions which sort of works but sometimes you want to suppress an annotation and walking them is inconvenient.
So now we have a list of them like usual.  One gotcha, to do previous schema validation we want them in order, so in that case we reverse the list early. We have to notice
this and not reverse it again later when semantic validation finishes.  But otherwise we get a nice clean list of annotations.

This lets us delete some janky code in other places and it will make the JSON the same as the other cases.

Two missing validations were added:

* any new unsub/resub directives must happen at version number >= any existing version number like all other directives
* deleted tables by construction do not need their parent tables anymore, once a table becomes  deleted we have to not verify the parents during `resub`, there is a lengthy comment on this in the code to clarify the situation, it's quite interesting.

Additionally, the subscriptions management was put under its own CRC at each version for additional resilience.  Restating past unsubscribes is really a horrible idea (hence resub) but if they are mixed with other upgrades a normal/safe addition to a past schema version (the only supported one is adding a new region to the upgrade) might disturb the a version with unsubscribes causing drops to re-run.  The new separation of versions is considerably more resilient.  You can only cause drops to re-run by changing subscriptions in the past which you aren't allowed to do.

Reviewed By: winniequinn

Differential Revision: D35200230

fbshipit-source-id: 54c972441c0d76fe1c0992303674092b3843e67f