CQL Guide Chapter 13 (JSON) rewritten
Summary:
This chapter was in need of overhaul.  The latest featuers have been added and it was rewritten
so that it is highly analogous to the JSON schema.  This should make it a lot easier to slot in
updates as they happen in the future.

Also regenerated the docs including two missing JSON replacements.

Differential Revision: D35413460

fbshipit-source-id: cb1f832ebe1e3740d40d1f52254564d1f2868c14