Include kinds in proc declare/create comparisons
Summary:
It was the case that a `CREATE PROC` definition would not be considered equivalent to its corresponding `DECLARE PROC` declaration if the former involved kinds. The only workaround was to omit the kinds from the declaration.

This fixes things so that kinds are not dropped when doing the comparison (and thus they must now be included in the declaration).

Reviewed By: ricomariani

Differential Revision: D35622322

fbshipit-source-id: fe617b57a0ecb511ec436d7055b41b60a0f26e8b