Allow (re)declaring and then creating OUT procs
Summary:
It was the case that one could not declare an OUT proc and then create it (or redeclare it) due to duplicate struct definitions being emitted which would cause C compilation to fail. For example, the C code emitted for this simple case could not be compiled:

```
declare proc p() out (x int) using transaction;

create proc p()
begin
  declare c cursor for select nullable(1) x;
  fetch c;
  out c;
end;
```

This fixes the problem by generating appropriate preprocessor directives to guard the declarations.

Reviewed By: ricomariani

Differential Revision: D35726707

fbshipit-source-id: 794bad4552101b5fc78ae38a4d2acc8ce9445869