If a table is unsubscribed we don't want to try to create any of the indices/triggers it might have
Summary:
We were erroneously trying to create indices and triggers on a table that was unsubscribed, this leads us to a bad place.

The correct thing to do is to blow away the index on unsubscribe and remove its CRC.

Triggers are even easier since we always drop and add them.  We just drop and never add if unsubscribed.

Many additional tests were added to force indices and triggers in all the various subscription voyage cases.

Reviewed By: winniequinn

Differential Revision: D35788105

fbshipit-source-id: 4b2ab4b0241830c8082436a70c8abf9f6018c1e0