Generalize CQL's internal hash table so that it can hash other kinds of keys and values
Summary:
This was previously used only to cache facets in the schema upgrade path but it will soon be used for other things too.

This is just a preparatory step.  Now you tell it how to hash and compare retain and release as needed.  It just does the hash table stuff.

The only convention is that the zero value has to be ok for empty.  Otherwise you get two cql_int64s to do with what you will.  It doesn't want to know or care if they are pointers.

(Note: this ignores all push blocking failures!)

Reviewed By: Strafos

Differential Revision: D35948403

fbshipit-source-id: 101b1b7d90444524b0965d1aecd35b9ce5049dbd