Add code generation for dynamic cursor calls
Summary:
We've introduced he notion of making an external function call with a generic cursor type.  The argument provided to the external function is a general shape description of the cursor, the same kind of shape descriptor that is used for the internal helpers like row fetching and so forth.  In fact the blob serialization helpers were redone to conform to this pattern, it's much simpler and there are fewer visible external pieces of data.

(Note: this ignores all push blocking failures!)

Reviewed By: Strafos

Differential Revision: D35858289

fbshipit-source-id: c230b244d93de25348a89272da973ca4ba8d501d