Additional error checks for unsubscription: make sure no views or triggers are using the table
Summary:
We have to note the tables used by any given view and the body of a trigger and use them just like we do with foreign keys during unsubscription.  The idea here is that if there is a live (not deleted) view or trigger that uses the table you can't unsubscribe to the table because you'll have errors.

Note that at this point you can't unsubscribe from views so that means any table used by a view is locked in. This will be added later.  But for now it's safe and will stay that way.

Differential Revision: D36178244

fbshipit-source-id: 51eb00a2cdb55d9b674820f28e65029445651930