Detect use of in arguments for an in/out call arg, adjust ref count accordingly
Summary:
If an `IN` arg `x` is SET we rename the arg from `x` to `_in__x`, then we copy it to a local `x` and we manage its lifetime.

We have to do the same thing for args that are set via a call with that argument in an `INOUT` our `OUT` position.  Same thing goes if the arg is set via `FETCH INTO`.

In this diff we just set the WAS_SET flag in those two additional places.  The codegen already knows what to do with the flag.

The actual fix is super simple, this diff is basically all tests.

Differential Revision: D36291979

fbshipit-source-id: 97f17e3b907697f26d3cd0bd5a0cda82e679f91a