Allow cql_extract_partition to return the same partition more than once
Summary:
This will happen in the parent/child case if the parent includes the same child key more than one time and this is perfectly reasonable in some cases.  The pointer is shared.

We use a bit stealing technique to tell if we have already created the result set.  If the LSB is 1 then the value is actually a result set pointer, otherwise it's a raw buffer like it was before.  Teardown logic to release the pointers when the partition is released was added.

FWIW, this is all pretty trivial stuff.

Differential Revision: D36384521

fbshipit-source-id: 003461283cec039736cb7eb2acde69ca98cd3ba3