disallow use of shared fragments with declare proc
Summary:
It's possible to put attributes anywhere however the semantic analyzer was not prepared to find `attribute(cql:shared_fragment)` on a proc *declaration*.

In order to actually use a shared fragment the compiler has to inline the body of it into the generated SQL.  So declares make no sense for them.  They are kind of like C++ inline functions in this regard.

However, this case wasn't detected so there was a crash when it tried to decode the AST.  There is now an error message for this case.

Repro:

```
attribute(cql:shared_fragment)
declare proc x() (x integer);

create proc y()
begin
  with (call x())
  select * from x;
end;
```

This now generates:

```
x:7:1: error: in str : CQL0468: attribute(cql:shared_fragment) may only be placed on a CREATE PROC statement 'x'
semantic errors present; no code gen.
```

Differential Revision: D36388289

fbshipit-source-id: 1c3125c11b157b9d6310e9163a42d8071713063f