add virtual tables to CQL schema generator integration tests
Summary: Adds virtual tables to V3 and V4 of the schema used for schema generation tests. This diff unconditionally deploys all tables, but the next diff will *not* deploy virtual tables in V3, then deploy *all* tables in V4. This should be sufficient to test the upgrader when using the no virtual tables deployment sproc.

Differential Revision: D35958565

fbshipit-source-id: 9dcaa1dd79be5d6a2d6a3bd5d1cc57d54d9367fd