Update "Getting Started" docs
Summary:
Updated introductory CG/SQL documentation so it might be a bit easier to follow for newcomers. I mainly removed the duplication between the GitHub Readme doc and the "Getting started" section in the website, and directed users to the "Hello World" program in Chapter 1 after they build `cql`.

Reorganized the Quick Start section so we differentiate pages about using CG/SQL vs developing for it.

I made some changes to the "Hello World" section to improve the flow - I felt the point about including cqlrt should be brought up earlier, and reduced some duplication by linking to existing documentation.

Removed indentation in code blocks that affect the actual output of the document.

Reviewed By: ricomariani

Differential Revision: D36382309

fbshipit-source-id: 6de7abf25ed135e1792a73f95759785343429564