Fixed a minor problem in the parent/child rewrite allowing args to flow
Summary:
The call statement node was not used in either the parent or child call instead the code was consing up a new call_stmt node which is just wrong.  The whole point of the syntax was to allow the node to be trivially copied.

Fixed and added a semantic test.

Differential Revision: D36780212

fbshipit-source-id: 5cbc464537ff3f4c99fb3e58e10c65157df72ef9