Use result set types for child set columns to save casts
Summary:
All the getters and setters were just using the vanilla cql_object_ref type for result set columns.  That requires some boilerplate casts all over.  Here we make the public interface use a more precise type.

Secondarily, when I attempted to add the above type precision to the optional result set setters I discovered that setters were not being generated at all for nullable primitive types which is a clear oversight.  However, then I discovered that the low level helpers for setting the primitives did not match the public contract of the result set types, especially in the "type based getters" codegen.  This was previously not a problem because the broken cases weren't being generated at all.

I redid the setters contract so that it basically exactly matches the getters contract which allowed for a good bit of deletion.  There are test cases for all of this now.

Reviewed By: wanaya

Differential Revision: D36979549

fbshipit-source-id: 2dd8c5eccfcabe4447c106f4a8f267bef6f0ccf6