Resolve gcovr issues by using gcovr directly
Summary:
I've applied the workaround in on a full-time basis:

https://github.com/facebookincubator/CG-SQL/issues/92

To do this I had to use gcov directly rather than just gcovr.  This is not really a big deal because all our outputs are in one directory anyway.  But to make this work I had to make sure the Makefile was being more careful about not re-using targets.

* the compat version of strlike was making things needlessly complicated and we don't need it anymore, even the oldest things we want to support have sqlite3_strlike now, got rid of that
* cleaned up Makefile_common dependencies so that the same target is not built two ways
* used clqrt_mockable throughout test.sh so that coverage doesn't have to mix the mockable and not mockable version
  * I had to add a pass through mock in the upgrader tests, no big deal there

The main point of all of this is to allow us to do `gcov out/*.gcda` then add `-g -k` to `gcovr`  which causes gcovr to get manually created .gcov files avoiding Issue #92

cqlrt_cf had been broken for a bit, needed a minor fix.  Remove the compat version from there too.

Reviewed By: RaoulFoaleng

Differential Revision: D37526120

fbshipit-source-id: 66631841e0962a40c8d4e1b5b6f30b08dc7e2822