In upgrade test: Wrap the upgrade process in a transaction as is normally done. (#131)
Summary:
This has the nice side effect of improving the speed of the upgrade test considerably.
Write diagnostics for fails consistently to stderr.
Added an option to include verbose error tracing in case of upgrader failure.

Note that the upgrader is not affected by this, it's still up to the caller to decide if/when to use a transaction.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/131

Reviewed By: RaoulFoaleng

Differential Revision: D37585080

Pulled By: ricomariani

fbshipit-source-id: c69ec0c2e02265e804d4ea7ae97ec428af0d1e0b