allow doc comment for columns
Summary: Here we plumb the doc comments concept in for columns also.  This causes us a bit of consternation because the table comment will be lost by the time we need it.  The comment can only be added when the node is reduced.  With column comments supported the table comment would be clobbered by the column comment.  To work around this we add a reduction much earlier in the table flow so that we can save the comment to use later.

Differential Revision: D37765253

fbshipit-source-id: a5628e4b98ba994f273798da2b2fffa41d0b768a