LUA codegen initial checkin -- highly experimental
Summary:
This diff includes  a lot of code to generate LUA from CQL.

This is all very provisional, and the ABI might change; in particular handling of cursors is likely to change and indeed this whole business might disappear.  But, by landing, this at least it will be in the history in the event that someone wants it.

cqlrt.lua for instance assumes lsqlite3 and sqlite3 installed by brew which don't seem like sustainable long term choices.

The rest of the system is largely untouched.

NOTE: cg_lua.c was forked from cg_c.c and simplified tremendously but it has not had all its comments changed.  Most of the comments are a lie.  This will be fixed.

Reviewed By: winniequinn

Differential Revision: D33522323

fbshipit-source-id: cbede89d6ba4fa1585164fb81f5986a37ee47904