Resolve named types when comparing declarations
Summary:
It was the case that named types were not always resolved before comparing declarations. In particular, `sem_declare_group_stmt` would fail in situations like the following:

```
declare t type int;

declare group g
begin
  declare x t;
end;

declare group g
begin
  declare x t;
end;
```

In the example above, first, the declaration `g` would be checked. When analyzing `declare x t;`, `t` would be resolved to `int`. After that, the redeclaration of `g` would be checked. Since `g` is a redeclaration, the text for both the original declaration *post-analysis* and the redeclaration *pre-analysis* would be compared. This comparison would fail because `t` was never resolved to `int` in the redeclaration.

The reason `sem_declare_group_stmt` compared ASTs before analyzing the redeclaration is that the redeclaration *cannot* be analyzed: Doing so would result in duplicate variable name errors. Other statement types (e.g., proc declarations) do not have the same issue.

The fix is to add a new callback to "gen_sql.c"  that allows for named type resolution. This approach is cleaner than the alternative: `sem_declare_group_stmt` would otherwise need to extract and analyze all of the types and expressions within it before doing the comparison. Having "gen_sql.c" call out to do the resolution is a simpler, more general approach that will just work as new types of re-declarable statements are added to the language.

Reviewed By: ricomariani

Differential Revision: D37938633

fbshipit-source-id: ebb5a48858058aeafc9289c107375b6e91d6d059