regenerated cql tester
Summary:
This tool is used to verify cql outputs.  It is made by CQL so we usually check it in rather than have the tester depend on the very latest (maybe broken) CQL.

This is just a refresh.  In this case you can see the new builtin function CQL supports and light formatting differences.  There are no semantic changes at all.

Reviewed By: RaoulFoaleng

Differential Revision: D38105613

fbshipit-source-id: 143cab935a4070e0a7d73caeacdbb12c4064c139