Dynamic cursor structure includes field names
Summary:
adding field names to the dynamic cursor makes it useful for lots of diagnostic purposes.  The complex cql_cursor_format rewrite can be replaced with a small helper function.

Note that the string literals between various fields/cursors can be folded at link time so the cost is lower than it seems.

Reviewed By: winniequinn

Differential Revision: D38105391

fbshipit-source-id: 1dda1efd8249e77d01f4c3da438bc81c6d14e6eb