When calling functions to get a result set we have to release the out arg
Summary:
This is normal call stuff but we missed the new case where we're calling a function that returns a result set.

The call result always goes into a temporary, that temporary might have other stuff in it from a previous call, maybe in a loop, maybe elsewhere.  We have to free it.  Note that the function we call has to assume that the out arg might have JUNK so it can't free it.  The ABI is that the caller does so if it is non-junk.

Differential Revision: D38170857

fbshipit-source-id: 34088ffe6647be64a5c13f4f03b664ffb7fc2d33