Allow explicit creation of compressed strings if things are highly duplicated
Summary: This leverages the string pieces code that we use for --compress in a general way so that if you write  cql_compressed("foo") then you'll get the CQL compression strategy on that text in this module.  This is not very useful unless the strings are redundant with each other.  If --compress is not enabled then you get a normal string literal.

Differential Revision: D38220238

fbshipit-source-id: 5bdffd87b8cec7e229fb70c9406d73a777f388de