If a proc decl we need requires another proc, emit them recursively (#137)
Summary:
The point of all of this is so that if you have a procedure that is doing parent/child result set stuff and it has the child proc in say the out position its exported declaration will look something like this:

```
 declare proc foo() (my_child object<child_proc set>);
```

This is all fine and well but child_proc is not defined so that decl will fail.

Instead of demanding that users put the right headers in exactly the right order we recursively add anything you need.  So in the above you'll get something like:

```
 declare proc child_proc() (id integer, name text);
 declare proc foo() (my_child object<child_proc set>);
```

It's just a little recursion and buffer swizzling.

test helpers would have had the same problem so they get the help too.

And we don't want the same proc more than once so we keep a little symbol table of what we've already emitted.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/137

Test Plan:
tests added

./cov.sh

Reviewed By: winniequinn

Differential Revision: D38233548

Pulled By: ricomariani

fbshipit-source-id: de5278a14c1e7db8300ba66506403e0976e19085