if an int/long is passed as an arg to a proc or func expecting a bool it should be normalized
Summary:
This was previously done only for nullables, the theory being that passing say "10"
still results in a truthy bool but it creates a lot of confusion and inconsistency.

We normalize in all call cases now, just like stores.  Generally a proc call and arg
bind is kind of like an assignment so it makes sense to use the same rules.

Differential Revision: D39145722

fbshipit-source-id: f73e458b84841f92449c83febb264ee181cbff1f