Allow a procedure to support an interface regardless of what order the interface columns appear
Summary:
We ultimately want to support procs with more than one interface.  The first step is to allow the procedure to be a superset of the required columns.
You just have to have the required columns *somewhere*.  In the next step we will look for additional interface implementations and verify that all columns from all interfaces are present somewhere (and are compatible).

The original test cases were a bit weak in terms of error checks so I shored them up too.

Differential Revision: D39211906

fbshipit-source-id: 8de7e5eea3292e9b7f827a6c9ae3eeeeec38590a