Add support for multiple interfaces on a procedure
Summary:
The existing code just needed a minor tweak to validate all interfaces on a procedure.  With that done you can annotate for as many interfaces as you like.
Note that this causes no direct code-gen, only validation and output in the JSON.  But JSON based codegen can assume interfaces have been validated which is the point.

Differential Revision: D39216338

fbshipit-source-id: 24191653c9ae032e770aa44984c7c7764d7a193f