introducing cql_blob_get primitive (semantic checks only)
Summary:
We'll be adding more features to transparently get fields out of blobs.  The idea here is that you can use the standard helper cql_blob_get and it will be remapped to an extracting UDF of your choice which gets a standard arg signature including the blob and the hash of the field/type that it wants to extract.

For now this is just semantic checking of the cql_blob_get call, there is no codegen for this so using it until its mapped will net you errors and nothing else.  More diffs coming soon.

Usage:

```
cql_blob_get(a_blob, backed_table.backed_column)
```

this will compile into:

```
your_extracting_udf(a_blob, 123456)
```

Where 123456 is the hash of `backed_column` and its type.

Configuration features coming....

Reviewed By: RaoulFoaleng

Differential Revision: D39485448

fbshipit-source-id: c566ad7c04042f33748ebed87a54dbac960d4126