Add SHA256 helper and test cases for it
Summary:
We'll need this to generate hash codes for fields in our blob backed storage

Other system have standardized on SHA256 rather than the 64 bit CRC we use elsewhere so we're including this.
It's a tiny amount of code in any case.

The SHA256 code is public domain code due to Brad Conte (brad AT bradconte.com).  Slightly modified to build in this directory.

Differential Revision: D39522788

fbshipit-source-id: f37ac1a0d8019d0040e8a57889e709c7f1313ad0