rewrite cql_blob_get to be the key or value blob getter, use hash codes or offsets as needed
Summary:
We can finally do something useful with all this blob stuff. cql_blob_get is rewritten to the specified function and either uses a hash code or not.
The hash code is a combination of the field name and the field type passed through SHA256.  It's not super fast but it's highly portable.

Differential Revision: D39559714

fbshipit-source-id: 4bc9e9b6b5eae43f1d1dcc39f3bf51e153b3c0d4