Automatically create a shared fragment to read from the backing table of a backed table
Summary: When using a backing table we automatically create a shared fragment that can access it.  Next we will rewrite the select users to use this fragment but for now you can test it by using  (call _backed()) in place of backed.  The key/value mapping is done automatically.  This leaves only the replacement to do for the select case.

Differential Revision: D39597111

fbshipit-source-id: 1f58a74265bd7a42b8350dbb3bee8ebc3b37ebca