minor correction for a compiler warning in sha256, and simplify body of cursor ref reset
Summary:
Having trouble creating on a warning-less version of sha256 for both clang and gcc.  Cleaned this up with and additional cast.

The cqlrt_common function to reset a cursor before we deserialize into it can use the standard row cleanup function.  When it was first written the ref count and offset wasn't in the dyn cursor but it was added and so it can be simplified.

These are both no-op changes.

Reviewed By: RaoulFoaleng

Differential Revision: D39638289

fbshipit-source-id: 8d65fbe3942612cc9be22743e046f5fee5760553