Complete backing and backed table validations
Summary:
Backing and backed tables have to have a least one pk field and at least one non-pk field for the key and value blobs respectively.
Primary key fields detected by primary key constraint now get a flag on them "partial pk" so that they can be found easily to make the above checks easy.
Backing tables must have exactly two columns both of which are of type blob.
Addition of partial_pk flag disturbed some test output but this was easily corrected (only 9 cases were affected and only trivially).

The helpers to get the column index of a pk or a value can be re-written to use the new partial_pk flag making this computation also easier but that is not in this diff.  Look for that soon.

These validations were the only ones we had not yet enforced (that we know of).

Reviewed By: paulvanderspek

Differential Revision: D39670681

fbshipit-source-id: 7c24fad863c7b404a3225d4eaf62200832892af6