Change unsubscription strategy a little for better resilience
Summary:
If the baseline schema is disturbed then unsubscribed tables might be created again.  This diff puts the unsubscription
on the same plan as normal deletes which probably they should have been in the first place.  Delete has this same issue,
a deleted table can come back in the baseline and not dirty the version that it was deleted in.  We do two things:

* when we resub, we drop first in case there is a vestigial baseline version of the table that just got recreated
  * at that time we also force the indices to be recreated just like we used to in the drop logic
* when we unsub, we do nothing but note that there is an unsub
* at the end, any tables marked to unsub are deleted in drop order (reverse of create order)

This puts unsub on the same plan as delete and we're much more resilient to inadvertent schema changes

Differential Revision: D39706380

fbshipit-source-id: 3d99111b3bbdb3619cd25fe8522a4cd66d23cc6b