Add rowid as an explict field available in backed tables
Summary:
Note this is going to be less perfect than the normal rowid because it will tend to leak into things as it looks like an explicit column to both CQL and SQLite.  The rowid column is normally invisible if you don't mention it.  I may have to change the select * expanding logic so that it doesn't ever include columns named rowid.

The rowid is generally useful for queries and it is needed in order to do backed deletes economically (see next diff)

Reviewed By: paulvanderspek

Differential Revision: D39778099

fbshipit-source-id: 7b3d758c8c2aaac5eb910879c009310002aa73da