canonicalize arg order in blob functions, add lua helpers (#143)
Summary:
We standardize on the order:

hashcode/offset, value, type in all the functions

* sometimes there is no hash/offset because it's implied (e..g. create key in offset mode)
* sometimes there is no type because it's implied (all key update functions)

But the args always appear in this order if they are present.

The lua run test doesn't really work because the luasqlite3 library gives you no way to return an int64 from a UDF losslessly so we can't do type hashes.  I hacked a version that did 32 bit type hashes and it passes.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/143

Differential Revision: D39879236

fbshipit-source-id: 63baed887ca61334385166e8ae5c74680229dfab