allow creation of shapes by removing columns from a previous shape
Summary:
A long-standing need was to be able to create new shapes using existing shapes minus some columns.  A classic example of this is the need to create a shape from a table excluding its primary key, or maybe excluding its auto-increment field.

There are many use cases for this kind of thing but this is the canonical example:

```
create table foo(
  id long primary key autoincrement,
  x text,
  y text,
  z real
);

create proc insert_into_foo(cols like foo(-id), out id long not null)
begin
  insert into foo values (NULL, from cols);
  set id := get_last_insert_rowid();
end;
```

This can save you a lot of typing if `foo` has a lot of columns

The essential operation is `like foo(-id)` meaning that I want the columns of `foo` but not the `id` column.

Rules:

* you can't use -x to remove all the columns
* you can't mix x and -x in the same operation
  * you could use an intermediate shape to add some and remove others if that were necessary, or if you wanted one with the id and one without

To do this work we generalize the shape_def node so that takes "shape expressions" rather than just a name list.  The shape expressions for now are either 'x' or '-x'.  You could imagine generalizing this to "all the not null fields" or "all the fields that start with q" or whatever.

Reviewed By: RaoulFoaleng

Differential Revision: D40069568

fbshipit-source-id: d426f37d3019137ffa6c6f68a2b9dbf55e26ffd9