simplify handling of UDFs that take an object
Summary:
We added this ptr(x) business to allow you declare a SQL function that takes an int but really is going to get an object.  This was a workaround for the fact that SQL functions did not allow object types in their signature.  We had to fix that when we added table valued functions and the whole ptr() thing became somewhat silly.  But it's still there.  It's a no-op.

This diff removes the need for it entirely, handling all UDFs the same not just TVFs.

I ran int this trying to add object binding in the `--rt lua` path which was missing to this point.  The lua bit is actually really simple and mostly just needs some test code and a couple of weak tables.

Both things are added and the tests all pass now including the lua versions.  There are test implementations of "rscol" and "rscount" for lua too now so those tests can be enabled.

Ptr can be turned into a total no-op after this, it already nearly is...

Reviewed By: RaoulFoaleng

Differential Revision: D40187166

fbshipit-source-id: 85e62b44fe1818dd9af4877fff7c108adc0c8baa