Do not use CRC to decide if columns and tables need to be created
Summary:
We have a handy hash table that we established previously, use that instead

When we first started on this path we had very little in-memory state for the upgrader and it would look directly at `sqlite_master` for each potential column upgrade.  However, a while back we put in an optimization where we pre-read `sqlite_master` into a string dictionary.  This lets us check for columns without having to prepare a statement every time.

However, now that we have this handy dictionary there isn't really any need to use the CRCs to avoid upgrade steps.  We can simply check each one directly avoiding a lot of CRC lookups and possible error states in favor of just looking for tables and columns directly in the dictionary.  We have the dictionary anyway and the lookups are only a few hundred nanoseconds a pop...  The CRC check isn't much faster than that.

This also avoids any weird states which we've seen from time to time where the CRC indicates that an upgrade step ran but the table got reset to an older state some how.  The new system will simply notice that columns are missing and add them.

Differential Revision: D40197493

fbshipit-source-id: 1c416a109e363526c4aa53e0414c7d4b9a054544