When we are about to move a table from create to recreate we need to clean up old tables
Summary:
Both resub and create from recreate suffer from the situation that there might be an existing version of the table that is old.

For the create from recreate case you could imagine an arbitrarily different version of the table that was the last recreate version.  We can't upgrade in the normal way from there to the current version.  We have to do one last "recreate".  So we need to drop whatever is there before we make the new version that's on the create plan.

That's all fine and well but actually we might have to drop LOTS of tables.  The existing table might have any number of foreign keys.  So what we do is walk the foreign key chain and drop everything.  This is safe to do because if the current able has any foreign keys and it is on the create plan that all other tables that depend on it were also previously on the recreate plan and we are going to need to nuke the lot in this iteration.

So we do a recursive purge one time before making the new version.

We are in exactly the same situation when a resubscribe happens.  However, the situation is much easier here.  When the resub happens we just re-enable all of the alter table add column functions as well as the baseline schema if any.  If there is a stale version it will be upgraded as usual as the alter table add columns reappear.  So it just works!

Differential Revision: D40243770

fbshipit-source-id: 2408642e7c5d80b067aa09675a0f4823ab43558e