Update Extension with latest CQL changes
Summary:
Update Extension with latest CQL changes.

The Tree-sitter parser for CQL was broken because of `create_table_prefix_opt_temp` rule with the message error **`Bad pattern structure at offset 1102: '(CREATE)'...`**. Which shouldn't actually be a rule. But we added it to be able to grab the doc comment in create table statement.
https://www.internalfb.com/code/fbsource/[a792ee08b0ac]/xplat/msys/tools/cql/cql.y?lines=598-603

This diff update `tree_sitter.py` to suppress un-necessary rule `create_table_prefix_opt_temp`

Reviewed By: ricomariani

Differential Revision: D40278569

fbshipit-source-id: 163c5046cc5a93371061d3a6405334c759bc4950