Change prefix of emitted C functions if a proper attribute is set
Summary:
We use "cql:alt_prefix" as the attribute that override the the "symbol_prefix" provided by RT if any. It only overrides the main c function name, and not its helper functions such as getters.

One use case of the prefix override is when CQL user wants to provide a custom implementation of a proc in C while keeping all helper functions the same, and without changing code that calling the CQL-C functions.

Reviewed By: ricomariani

Differential Revision: D40331702

fbshipit-source-id: a0411008e1db9ed83c236a73850fe2fce6f36463