if parent recreate group recreates all children recreate
Summary: In this diff we emit schema upgrader code that checks whether a recreate group was recreated or rebuilt. If it recreated, then all of the child recreate groups also need to recreate. We emit DROPs for tables in these child recreate groups. When the child recreate groups go on the recreate plan, there will be no table in master, so we will automatically recreate and not go through the rebuild path.

Reviewed By: ricomariani

Differential Revision: D39706544

fbshipit-source-id: 6f9b72709fcb551c1904b161056391d63620f5fd