Change recreate group order based on dependencies
Summary: When allowing recreate groups to depend on other recreate groups, we need to walk the groups in a different order than what they are initialized in. In this diff we introduce a group_ordinal for each recreate_annotation (which is based on the group creation order). We use this as the primary sort key when ordering recreate tables.

Reviewed By: ricomariani

Differential Revision: D39682477

fbshipit-source-id: 9f169a41fc97d399e3b18a4d63f1fe02300254b9