Store recreate group dependencies in sem.c
Summary:
This stack deals with creating a recursive hierarchy for recreate groups so that recreate groups can FK to other recreate groups.

In this diff, we introduce a name_list - which is a linked list of strings. We will use this as the value of a symbol table to keep track of all the recreate groups a recreate group FKs to.

We create a symbol table [recreate group name -> list of its foreign key dependencies]. We update this table in the old CQL0060 error case (when a recreate group tries to FK to another recreate group).

~~An important assumption for this entire stack is: if a recreate group does not have a group name (i.e. "singleton" group) we treat the table name as the group name.~~ Tried to enforce this in D40256427 but not possible. Instead prepending t_ and g_ in front of table names/group names.

Reviewed By: ricomariani

Differential Revision: D39682492

fbshipit-source-id: d10df13bf7e51be81680f9b1c66314a3da3f6660